Move async functions into a custom util file.
